ci: change tool install action (#28)
Use the same approach as `veecle-os` to install tools. This is in
preparation to switch to tombi, which needs to be installed from Git.
Installing from Git currently isn't supported by the previous approach.

I considered reusing the action directly via `veecle-os/...`, but that
would require us to add `cargo-docs-rs` to the action in `veecle-os`.

Signed-off-by: Julian Schindel <julian.schindel@veecle.io>